Remove `RhymeNotFound`
If a line has a last word that does not have a rhyme,
a new line will be generated instead.

`is_instant_fail`, although no longer serves a purpose,
is still kept just in case it will be used for future implementation.

It will be removed if that is not the case.