Remove major blocking funcs and made other changes
- A `ThreadPoolExecutor` is used to to runnon-async
functions asynchronously
- `_get_sentence` facilitates the async portion of the
markov chain's `make_short_sentence`
- `LinePagination.paginate` is used for the `poem_embed`. This is useful
for when the poem is really long.
- Added `MakeShortSentenceRanOut` custom error
- When a rhyming sentence cannot be found, the `RhymingSentenceNotFound`
error is raised, and the information stored will allow the `send_markov_poem`
to be called once more.
- Should it fail, the user will be asked to type the command again.
This is because an error cannot be raised inside the process of an error
being handled.
- Thus, the `RhymingSentenceNotFound` class now contains class variables
- - `_get_rhyming_line` also has `error_info` as a parameter for the
`RhymingSentenceNotFound` error to get the information required.
- `_hangle_rhyming_sentence_not_found` is added to separate
message-sending logic from other components.
- `RhymeAPIUnresponsive` custom error added. If an API is unresponsive,
the bot will stop generating the poem and tell the user to try again
next time.
- `is_instant_fail` variable removed. The while-loop of `init_unit` to
ensure that a line has a rhyme removes the need for it.
- The `Cache` context manager checks if the `word` inside the `cache`
before checking if it has an empty rhyme set.
This fixes the bug where if an error is raised, the `word`
is not stored in the cache and an attempt is made to retrieve it.
- Added `command_wrapper` to separate the `ctx.typing()` and timeout logic
from the poem generation process
- `POEM_TIMEOUT` currently changed to 60 seconds, likely to be changed.
- Added a few more to the `templates`.
- Added missing f-string in `_get_rhyme_set`.
- `_get_rhyming_line` will no longer return a line if it already
exists in the poem.
- `_init_unit` may return a lines with a last word without rhymes
if the unit only appears once in the rhyme scheme.
- Added `send_markov_poem` to separate poem generation logic
from user command logic.
- Add `help` and `brief` parameters to `commands.command` decorator of
`poem`.
- `error.handled = True` added to all local error local handlers to
stop the global error handler from giving an error.
- Added embeded responses to `RhymeAPIUnresponsive` and
`MakeShortSentenceRanOut` custom errors.
- Whenever the bot retries making a poem, the user is pinged and
told that it might take a while.
- If a poem times out, the user is pinged and is told to try again.