Only output override channels & bot commands channel on whitelist error
Previously this would output all channels, and could result in an error.
This change ensures only the main bot channel & and any overridden channels
are shown in the embed. We do this before the categories block as
the categories kwarg itself is an override, so we want to include
those in any output.