Realigned to SirLancebot Structure The code for rattlesnakes's internal eval was aligned to Sir Lancebot's structure. It was mostly renaming rattlesnake to bot and changing how some of the imports were setup as. It also included changing the __init__.py to match the Sir Lancebot cog structure.
Additionally, the whitelist check has been significantly
simplified to only be a role check for the admin role.
The rattlesnake implementation had a more robust
`in_whitelist` decorator, so it may be worth investigating
adding that in if we see fit. For now, it's a simple
`with_role` decorator check.

The name of the cog file itself was changed
to include an underscore to sidestep what I think
was a namespace collision that would prevent
the setup function from properly running.