Add logic for preserving unloaded Cogs.
This commit makes a redis cache in `bot.bot` to store all the unloaded
extensions, on every unload and load this cache would be edited in the
desired manner, either removing it from it or adding it in the cache.

Now, coming to not loading the unloaded commands, on bot startup, just
after redis gets connected, the bot would run `bot.get_unloaded_extensions`
and get all the unloaded extensions from the cache. This extensions
would be skipped and not loaded, the skipping message would be logged.
Once all these extensions are loaded, we compare and check which
extensions are present in the cache and not the bot. This means this
cogs are either remove or renamed.

A message containing all these "extra" cogs are sent to #dev-alerts and
these cogs are removed from the cache.

Resolves: #705