Reformat unloads redis cache
Earlier it stored all the unloaded extensions in a single key named `"unloaded"` which wasn't a suitable way as we manually serialized the cog names by joining them with a pipe character, then set them to a single key unloaded in the cache.

This commit changes the format to `{"ext": "msg link"}`, so now each extension would have its separate item in the cache containing the message link as the value.

The message link has been chosen here as it provides information about why the extension for was unloaded at the first place.