Don't role check if running lovecalc on yourself
This skips the role check if the author is running the love calc command on themselves. We still want to check the the user running against has the role, along with if the author is running it against 2 other people.

i have also included info on how to get the lovefest role in the error embed.