CI: check for license compatibility
Port of https://github.com/python-discord/bot/pull/1779

---

Since our project is licensed under the MIT License, we can't be using any incompatible license-wise dependencies in our project, such as GPL.

This commit adds a step to the CI lint phase that will use pip-licenses to verify that all the installed packages are part of an ALLOWED_LICENSE variable. This variable is currently set to be every license we currently use.

We opted to use an allowlist instead of a denylist to make sure that new licenses are reviewed by hand and added to that variable.