Add support to query AoC results in respect of days and stars
From now on the AoC leaderboard command accepts a total of 2 optional arguments a day and star string (eg.: 1-2, for the second star of the first day) and a number of results they would like to see, with a total maximum of 15.

This commit also introduces a few minor fixes in the AoC helper.