More changes to Madlibs
So...a _lot_ has been done here.
Been fixing a lot with the code and the templates, the main problem was figuring out a problem with spacing that @Bluenix2 helped me out with a bunch. Also fixed some docstrings, and changed some of the logic that appends the words submitted by the user to the story.