fix: hanukkah command respects month boundaries
rewrote hanukkah to use datetime.strptime
left a helper method and some variables in order to allow extending
to use a cache in the future, rather than requesting the api every invoke

that is out of scope of this commit and pull, since the command is
currently broken.

I've only kept the same functionality, without trying to rewrite the
entire command.