Add `.aoc link` command
This new command will allow people to associate their Discord ID with
their Advent of Code name. This Redis Cache idea was taken from the
hacktoberfest stats command, which allows people to associate their
github username to then pull the correct stats.

This does not check if the name exists on the leaderboard and that is
intentional. Due to the cooldown on the leaderboard I don't want to rely
on that before someone can link their account. Additionally, someone
may change their display name on the Advent of Code side and I don't
think validation of it existing on the leaderboard gets us anything.

The usefulness of this function will not be apparent in this cog, but it
is necessary for something fun I'd like to do.