Remove latex cog and matplotlib
matplotlib, and its sub dependencies, caused a fresh install of an environment to take multiple minutes. As the latex cog is the only one that used it, and that is currently disabled, we have decided to remove it entirely.

Git gives us the benefit of being able to see deleted files. So whoever decides to implement latex again can use that for reference.