Clean Up Constants File
This commit moves the logging constants from the Client class to a new
Logging class.

Flake8 is renabled for constants, and the error it was disabled for was
specifically ignored. All new errors were fixed.

It also fixes a bug with __all__, which was trying to export a missing
symbol due to a missing comma.

Signed-off-by: Hassan Abouelela <hassan@hassanamr.com>