Disable AoC completionist task
This disabled the completionist task that checks the leaderboard for people who have 50 stars and gives out the role. Since the event is running, we are not keeping the session cookies up to date, so this is flooding #dev-log with errors.

This task should be altered in preparation for next event so that commenting out this line isn't required.