Add non-Christmas holiday reaction capabilities
Triggers for reactions are moved from a dict to to a 2-deep NamedTuple.
This is cleaner for multiple reactions, and allows for additional parameters,
such as months.

Message case checking is altered from `.lower` to `re.IGNORECASE`.
Not only does the regex flag short circuit, it doesn't modify
the string in question